functions/requirements.txt
google-genai
firebase-functions
firebase-admin
